+++ src/__tests__/components/Menu.spec.js
@@ -4,6 +4,12 @@ import React from "react";
 import Menu from "../../componentes/Menu";
 
 describe("Testes do menu", () => {
+    it("Deve haver o componente menu", () => {
+        const wrapper = shallow(<Menu />);
+
+        expect(wrapper.exists()).toBe(true);
+    });
+
     it("Deve renderizar o menu", () => {
         const wrapper = shallow(<Menu />);
